refactor: move Resource to @opentelemetry/sdk-base
Moves the Resource interface to the SDK level by creating a new packaged
called sdk-base. Also helps resolve the issue of shared objects between
the metrics and trace SDK.
