+++ packages/opentelemetry-sdk-base/package.json
@@ -1,0 +1,73 @@
+{
+  "name": "@opentelemetry/sdk-base",
+  "version": "0.0.1",
+  "description": "OpenTelemetry SDK base",
+  "main": "build/src/index.js",
+  "browser": {
+    "./src/platform/index.ts": "./src/platform/browser/index.ts"
+  },
+  "types": "build/src/index.d.ts",
+  "repository": "open-telemetry/opentelemetry-js",
+  "scripts": {
+    "test": "nyc ts-mocha -p tsconfig.json test/**/*.ts",
+    "test:browser": "karma start --single-run",
+    "tdd": "yarn tdd:node",
+    "tdd:node": "yarn test -- --watch-extensions ts --watch",
+    "tdd:browser": "karma start",
+    "codecov": "nyc report --reporter=json && codecov -f coverage/*.json -p ../../",
+    "clean": "rimraf build/*",
+    "check": "gts check",
+    "compile": "tsc -p .",
+    "fix": "gts fix"
+  },
+  "keywords": [
+    "opentelemetry",
+    "nodejs",
+    "tracing",
+    "profiling",
+    "metrics",
+    "stats"
+  ],
+  "author": "OpenTelemetry Authors",
+  "license": "Apache-2.0",
+  "engines": {
+    "node": ">=8.0.0"
+  },
+  "files": [
+    "build/src/**/*.js",
+    "build/src/**/*.d.ts",
+    "doc",
+    "LICENSE",
+    "README.md"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "devDependencies": {
+    "@types/mocha": "^5.2.5",
+    "@types/node": "^12.6.8",
+    "@types/sinon": "^7.0.13",
+    "@types/webpack-env": "1.13.9",
+    "codecov": "^3.1.0",
+    "gts": "^1.0.0",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-mocha": "^1.3.0",
+    "karma-spec-reporter": "^0.0.32",
+    "karma-webpack": "^4.0.2",
+    "mocha": "^6.1.0",
+    "nyc": "^14.1.1",
+    "sinon": "^7.3.2",
+    "rimraf": "^3.0.0",
+    "tslint-microsoft-contrib": "^6.2.0",
+    "tslint-consistent-codestyle": "^1.15.1",
+    "ts-loader": "^6.0.4",
+    "ts-mocha": "^6.0.0",
+    "ts-node": "^8.0.0",
+    "typescript": "^3.6.3",
+    "webpack": "^4.35.2"
+  },
+  "dependencies": {
+    "@opentelemetry/types": "^0.0.1"
+  }
+}
