+++ packages/opentelemetry-sdk-base/test/resources/resource.test.ts
